% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ratio_functions.R
\name{adj_d_ratio}
\alias{adj_d_ratio}
\title{adj_d_ratio}
\usage{
adj_d_ratio(input)
}
\arguments{
\item{input}{An inner d object}
}
\value{
A vector containing the adjusted d-ratio in dependency of
sample size.
}
\description{
adjusted d-ratio with respect to sample size.
}
\details{
This function computed the adjusted d_ratio with respect to sample size.
}
\author{
Julian Urban
}
