% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inner_MAGMA_functions.R
\name{build_value_matrix}
\alias{build_value_matrix}
\title{build_value_matrix}
\usage{
build_value_matrix(input_list, rep_element, name_ps = "distance_ps")
}
\arguments{
\item{input_list}{Data with PS}

\item{rep_element}{sample sizes per group}

\item{name_ps}{names of PS in data}
}
\value{
the input for distance estimation
}
\description{
prepares distance estimation.
}
\details{
This function uses the PS inputs to prepare them for distance estimation.
}
