% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGMA_function.R
\name{distance_estimator}
\alias{distance_estimator}
\title{distance_estimator}
\usage{
distance_estimator(data, means, variance, cores, inp = NULL)
}
\arguments{
\item{data}{A matrix containing all possible combinations.}

\item{means}{A matrix containing all row means of all possible matches.}

\item{variance}{A numeric indicating the variance of the propensity scores.}

\item{cores}{An integer defining the number of cores used for
parallel computation.}

\item{inp}{input parameter for parallel distance computation.}
}
\value{
A matrix of distance for each case of each possible match.
}
\description{
estimates distance in \code{\link{MAGMA}}.
}
\details{
This function is an inner function of \code{\link{MAGMA}}. It estimates the
distance of all possible matches.
}
\author{
Julian Urban
}
