% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_parametric_descs.R
\name{effect_nominal}
\alias{effect_nominal}
\title{effect_nominal}
\usage{
effect_nominal(Data, group, variable)
}
\arguments{
\item{Data}{A data set}

\item{group}{A character specifying the grouping variable}

\item{variable}{Variables for which the effect should be estimated}
}
\value{
A vector or matrix containing nominal effect sizes
}
\description{
Computes descriptive statistics and effect size for nominal data
}
\details{
This function computes descriptive statistics and effect size for nominal
data.
}
\author{
Julian Urban
}
