% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_parametric_descs.R
\name{effect_ordinal}
\alias{effect_ordinal}
\title{effect_ordinal}
\usage{
effect_ordinal(Data, group, variable)
}
\arguments{
\item{Data}{A data set}

\item{group}{A character specifying the grouping variable}

\item{variable}{Variables for which the effect should be estimated}
}
\value{
A vector or matrix containing ordinal effect sizes
}
\description{
Computes descriptive statistics and effect size for ordinal data
}
\details{
This function computes descriptive statistics and effect size for ordinal
data.
}
\author{
Julian Urban
}
