% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ratio_functions.R
\name{inner_d}
\alias{inner_d}
\title{inner_d}
\usage{
inner_d(da, gr, co, st, co_ord = NULL, co_nom = NULL)
}
\arguments{
\item{da}{Specifying the data frame or tibble with the data.}

\item{gr}{A character vector specifying the IVs.}

\item{co}{A character vector naming the DVs.}

\item{st}{A character naming the variable for iteratively inclusion}

\item{co_ord}{A character vector naming the ordinal DVs.}

\item{co_nom}{A character vector naming the nominal DVs.}
}
\value{
A list of length two. The first element is a matrix including all
pairwise effects. The second is a vector expressing d-ratio
in dependency of sample size.
}
\description{
d-ratio and pairwise Cohen's d with respect to sample size.
}
\details{
This function computed the d-ratio and all pairwise effects
with respect to sample size.
}
\author{
Julian Urban
}
