% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_g_functions.R
\name{J_group_size}
\alias{J_group_size}
\title{J_group_size}
\usage{
J_group_size(group_size)
}
\arguments{
\item{group_size}{A numeric defining the max sample size for which J
should be computed.}
}
\value{
A vector of J's in dependency of sample size.
}
\description{
inner function of \code{\link{mean_g_meta}}.
}
\details{
This function computes J over samples sizes necessary for Hedges' g.
}
\author{
Julian Urban
}
