% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_g_functions.R
\name{mean_g_meta}
\alias{mean_g_meta}
\title{mean_g_meta}
\usage{
mean_g_meta(input, number_groups)
}
\arguments{
\item{input}{An inner d object.}

\item{number_groups}{A numeric specifying the number of groups.}
}
\value{
A vector containing the mean g in dependency of sample size.
}
\description{
Mean standardized effect.
}
\details{
This function computes the mean effect. Method varies between two and
many group matchings.
}
\author{
Julian Urban
}
