% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pillai_functions.R
\name{Pillai_iterativ}
\alias{Pillai_iterativ}
\title{Pillai_iterativ}
\usage{
Pillai_iterativ(da, gr, co, st)
}
\arguments{
\item{da}{Specifying the data frame or tibble with the data.}

\item{gr}{A character vector specifying the IVs.}

\item{co}{A character vector naming the DVs.}

\item{st}{A character naming the variable for iteratively inclusion}
}
\value{
A vector containing Pillai's Trace in dependency of sample size. If two
grouping variables were specified, the output is a matrix containing
Pillai's Trace for both IVs and their ineraction.
}
\description{
Pillai's Trace with respect to sample size.
}
\details{
This function computes Pillai's Trace for increasing sample size.
}
\author{
Julian Urban
}
