% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_parametric_descs.R
\name{row_ordinal}
\alias{row_ordinal}
\title{row_ordinal}
\usage{
row_ordinal(Data, group, variable)
}
\arguments{
\item{Data}{A data set}

\item{group}{A character specifying the grouping variable}

\item{variable}{Variables for which the effect should be estimated}
}
\value{
A vector containing the adjusted d-ratio in dependency of
sample size.
}
\description{
Computes descriptive statistics and effect size for ordinal data
}
\details{
This function computes descriptive statistics and effect size for ordinal data.
}
\author{
Julian Urban
}
